% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{by_owner}
\alias{by_owner}
\title{Filter content by owner(s)}
\usage{
by_owner(content, username)
}
\arguments{
\item{content}{Content data frame from \code{content(...)}}

\item{username}{The username of the owner}
}
\value{
The filtered content data frame
}
\description{
Returns content items owned by the specified user(s) (by username)
}
