% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{by_owners}
\alias{by_owners}
\alias{by_owner}
\title{Filter content by owner(s)}
\usage{
by_owners(content, usernames)

by_owner(content, usernames)
}
\arguments{
\item{content}{Content data frame from \code{content(...)}}

\item{usernames}{The username of the owner, or a list of usernames if multiple}
}
\value{
The filtered content data frame
}
\description{
Returns content items owned by the specified user(s) (by username)
}
