% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{by_tag}
\alias{by_tag}
\title{Filter content by tag(s)}
\usage{
by_tag(content, tagname)
}
\arguments{
\item{content}{Content data frame from \code{content(...)}}

\item{tagname}{The name of the tag}
}
\value{
The filtered content data frame
}
\description{
Returns content items that have the specified tag name(s)
}
