% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{by_tags}
\alias{by_tags}
\alias{by_tag}
\title{Filter content by tag(s)}
\usage{
by_tags(content, tagnames)

by_tag(content, tagnames)
}
\arguments{
\item{content}{Content data frame from \code{content(...)}}

\item{tagnames}{The name of the tag, or a list of names if multiple}
}
\value{
The filtered content data frame
}
\description{
Returns content items that have the specified tag(s) (by tag name)
}
