% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectwidgets.R
\docType{package}
\name{connectwidgets}
\alias{connectwidgets}
\alias{connectwidgets-package}
\title{connectwidgets: Curate your content on RStudio Connect}
\description{
connectwidgets provides UI components to help publishers provide curated
organization of content on RStudio Connect.
}
\section{Connection}{

Set \code{CONNECT_SERVER} and \code{CONNECT_API_KEY} environment variables and
\code{\link[=connect]{connect()}}. Fetch a \code{tibble} listing server content with \code{\link[=content]{content()}}. Use
the \code{\link[=by_owner]{by_owner()}} and \code{\link[=by_tags]{by_tags()}} filter helpers to narrow your search.
}

\section{Components}{

Present your content with different components:
\itemize{
\item \code{\link[=rsc_card]{rsc_card()}} for highlighting individual content items
\item \code{\link[=rsc_grid]{rsc_grid()}} for groups of cards
\item \code{\link[=rsc_table]{rsc_table()}} for a tabular view
\item \code{\link[=rsc_search]{rsc_search()}} a text input for searching grids and tables
\item \code{\link[=rsc_filter]{rsc_filter()}} an input to filter grids or tables
by owner, type, or tag
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/connectwidgets/}
  \item \url{https://github.com/rstudio/connectwidgets}
}

}
\author{
\strong{Maintainer}: Brian Smith \email{brian@rstudio.com}

Authors:
\itemize{
  \item Marcos Navarro \email{marcos@rstudio.com}
}

Other contributors:
\itemize{
  \item David Aja \email{david@rstudio.com} [contributor]
  \item Kelly O'Briant \email{kelly.obriant@rstudio.com} [contributor]
  \item RStudio [copyright holder]
}

}
