% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content}
\alias{content}
\title{Get Content Items}
\usage{
content(client)
}
\arguments{
\item{client}{A Client object (see \code{connect})}
}
\value{
A tibble of content items
}
\description{
Returns content items from the Connect server as a tibble that are visible to the API key's user account
}
