% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_widget_input}
\alias{evaluate_widget_input}
\title{Evaluates required columns for widget's input}
\usage{
evaluate_widget_input(widget, colnames, required)
}
\arguments{
\item{widget}{Widget's name to identify errors thrown}

\item{colnames}{Column names provided from the widget content input}

\item{required}{List of required columns to look for in the widget's input}
}
\description{
Evaluates required columns for widget's input
}
