% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_bootswatch_theme_name}
\alias{get_bootswatch_theme_name}
\title{Get docmeta theme name (output: html_document: theme)}
\usage{
get_bootswatch_theme_name()
}
\description{
Get docmeta theme name (output: html_document: theme)
}
