% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_current_bootswatch_theme}
\alias{get_current_bootswatch_theme}
\title{Get the current bootswatch theme if any}
\usage{
get_current_bootswatch_theme()
}
\description{
Get the current bootswatch theme if any
}
