% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{get_current_theme}
\alias{get_current_theme}
\title{Verify if a theme is already set within document meta,
or within the user's code via bslib}
\usage{
get_current_theme()
}
\description{
Verify if a theme is already set within document meta,
or within the user's code via bslib
}
