% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_user_provided_theme}
\alias{get_user_provided_theme}
\title{Get the current user provided bslib theme if any}
\usage{
get_user_provided_theme()
}
\description{
Get the current user provided bslib theme if any
}
