% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{resolve_theme_dependency}
\alias{resolve_theme_dependency}
\title{Resolve and get theme to be used by a widget. It could be the
default connectwidgets styling theme or one provided by the user.}
\usage{
resolve_theme_dependency(widget_name)
}
\arguments{
\item{widget_name}{The name of the widget (e.g: rsc_grid)}
}
\description{
Resolve and get theme to be used by a widget. It could be the
default connectwidgets styling theme or one provided by the user.
}
