% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsc_card.R
\name{rsc_card}
\alias{rsc_card}
\title{Card view for content}
\usage{
rsc_card(content)
}
\arguments{
\item{content}{A data frame from Connect's content. Requires the
following columns "guid," "url", "title".
And, although optional, expects an "owner_username", "description"
and "updated_time" columns.}
}
\description{
Renders a card view for the provided content
}
