% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsc_search.R
\name{rsc_search}
\alias{rsc_search}
\title{Search widget}
\usage{
rsc_search(content)
}
\arguments{
\item{content}{A shared object from Connect's content}
}
\description{
Search for matching content name/title within the shared data object passed.
Expects the exact same frame passed to the view widget being filtered.
}
