% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsccard.R
\name{rsccard}
\alias{rsccard}
\title{Card view for content}
\usage{
rsccard(content)
}
\arguments{
\item{content}{A data frame from Connect's content. Requires the
following columns "guid," "url", "title".
And, although optional, expects an "owner_username", "description"
and "updated_time" columns.}
}
\description{
Renders a card view for the provided content
}
