% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscfilter.R
\name{rscfilter}
\alias{rscfilter}
\title{Filter widget}
\usage{
rscfilter(content)
}
\arguments{
\item{content}{A data frame from Connect's content. Although optional,
expects an "owner_username", "app_mode" and "tags" columns.}
}
\description{
Filter content rows with owner, content type and tags,
expects the exact same frame passed to the view widget being filtered.
}
