% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscpages.R
\docType{package}
\name{rscpages}
\alias{rscpages}
\alias{rscpages-package}
\title{rscpages: Curate your content on RStudio Connect}
\description{
rscpages provides UI components to help publishers provide curated
organization of content on RStudio Connect.
}
\section{Connection}{

Set \code{CONNECT_SERVER} and \code{CONNECT_API_KEY} environment variables and
\code{\link[=connect]{connect()}}. Fetch a \code{tibble} listing server content with \code{\link[=content]{content()}}. use
the \code{\link[=by_owner]{by_owner()}} and \code{\link[=by_tags]{by_tags()}} filter helpers to narrow your search.
}

\section{Components}{

Present your content with different components:
\itemize{
\item \code{\link[=rsccard]{rsccard()}} for highlighting individual content items
\item \code{\link[=rscgrid]{rscgrid()}} for groups of cards
\item \code{\link[=rsctable]{rsctable()}} for a tabular view
\item \code{\link[=rscsearch]{rscsearch()}} a text input for searching grids and tables
\item \code{\link[=rscfilter]{rscfilter()}} an input to filter grids or tables by owner, type, or tag
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/rscpages}
  \item \url{https://github.com/rstudio/rscpages}
}

}
\author{
\strong{Maintainer}: Brian Smith \email{brian@rstudio.com}

}
