% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscsearch.R
\name{rscsearch}
\alias{rscsearch}
\title{Search widget}
\usage{
rscsearch(content, width = NULL, height = NULL)
}
\arguments{
\item{content}{A shared object from Connect's content}

\item{width, height}{Optionally specified width and height of the widget}

\item{elementId}{Optional HTML id of the element}
}
\description{
Search for matching content name/title within the shared data object passed.
}
