% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warning_large_content}
\alias{warning_large_content}
\title{Show warnings for large content}
\usage{
warning_large_content(content, max_rows = 500)
}
\arguments{
\item{content}{Content data frame or Crosstalk SharedData object}

\item{max_rows}{Maximum number of rows before warning (default 500)}
}
\description{
Show warnings for large content
}
