% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{warning_widget_input}
\alias{warning_widget_input}
\title{Show warnings of expected columns for widget's input}
\usage{
warning_widget_input(widget, colnames, expected)
}
\arguments{
\item{widget}{Widget's name to identify warnings thrown}

\item{colnames}{Column names provided from the widget content input}

\item{expected}{List of expected columns to look for in the widget's input}
}
\description{
Show warnings of expected columns for widget's input
}
