% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Create a connection to RStudio Connect}
\usage{
connect(
  host = Sys.getenv("CONNECT_SERVER", NA_character_),
  api_key = Sys.getenv("CONNECT_API_KEY", NA_character_)
)
}
\arguments{
\item{host}{The host URL for accessing RStudio Connect. Defaults to environment
variable CONNECT_SERVER}

\item{api_key}{The API key to authenticate with RStudio Connect. Defaults
to environment variable CONNECT_API_KEY}
}
\value{
An RStudio Connect R6 object that can be passed along to methods
}
\description{
Creates a connection to RStudio Connect using the hostname and an api key.
}
