% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{content}
\alias{content}
\title{Get Content Items}
\usage{
content(connect)
}
\arguments{
\item{connect}{A Connect object}
}
\value{
A tibble of content items
}
\description{
Returns content items from the Connect server that are visible to the API key's user account
}
