% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{default_theme}
\alias{default_theme}
\title{Principal (most used) theming variables}
\usage{
default_theme()
}
\description{
bg
fg
primary
secondary
success
info
warning
danger
base_font
code_font
heading_font
font_scale
}
\details{
Visit https://rstudio.github.io/bslib/articles/bs4-variables.html
to see the full list of theming variables available
}
