% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{gen_theme_dependency}
\alias{gen_theme_dependency}
\title{Generate the theme's bslib::bs_dependency to be used by a widget.}
\usage{
gen_theme_dependency(widget_name, theme)
}
\arguments{
\item{widget_name}{The name of the widget (e.g: rscgrid)}

\item{theme}{The bslib theme to generate the CSS dependency}
}
\description{
Generate the theme's bslib::bs_dependency to be used by a widget.
}
