% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{gen_theme_dependency}
\alias{gen_theme_dependency}
\title{Generate the theme's bslib::bs_dependency to be used by a widget.}
\usage{
gen_theme_dependency(widget_name, theme, default_base = FALSE)
}
\arguments{
\item{widget_name}{The name of the widget (e.g: rsc_grid)}

\item{theme}{The bslib theme to generate the CSS dependency}

\item{default_base}{Using the default theme or not}
}
\description{
Generate the theme's bslib::bs_dependency to be used by a widget.
}
