% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsccard.R
\name{rsccard}
\alias{rsccard}
\title{Grid view for content}
\usage{
rsccard(client, content, width = NULL, height = NULL)
}
\arguments{
\item{content}{A shared object from Connect's content}

\item{width, height}{Optionally specified width and height of the widget}
}
\description{
Renders a grid view of the provided content items
}
