% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscfilter.R
\name{rscfilter}
\alias{rscfilter}
\title{Filter widget}
\usage{
rscfilter(content)
}
\arguments{
\item{content}{A shared object from Connect's content}
}
\description{
Filter content rows with owner, content type and tags,
from a shared data from Connect's content.
}
