% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{rscfilter}
\alias{rscfilter}
\title{Placeholder for the custom filtering htmlwidget}
\usage{
rscfilter(content, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{content}{The tibble of content provided by rscpages::content()}

\item{width, height}{Optionally specified width and height of the widget}

\item{elementId}{Optional HTML id of the element}
}
\description{
<Add Description>
}
