% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscgrid.R
\name{rscgrid}
\alias{rscgrid}
\title{Grid view for content}
\usage{
rscgrid(content)
}
\arguments{
\item{content}{A data frame from Connect's content. Requires the
following columns "guid", "url", "title", "app_mode", "owner_username".
And, although optional, expects an "updated_time" column.}
}
\description{
Renders a grid view of the provided content items
}
