% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscgrid.R
\name{rscgrid}
\alias{rscgrid}
\title{Grid view for content}
\usage{
rscgrid(content)
}
\arguments{
\item{content}{A shared object from Connect's content}
}
\description{
Renders a grid view of the provided content items
}
