% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscsearch.R
\name{rscsearch}
\alias{rscsearch}
\title{Search widget}
\usage{
rscsearch(content)
}
\arguments{
\item{content}{A shared object from Connect's content}
}
\description{
Search for matching content name/title within the shared data object passed.
}
