% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsctable.R
\name{rsctable}
\alias{rsctable}
\title{Reactable table of the content}
\usage{
rsctable(content)
}
\arguments{
\item{content}{The tibble of content provided by rscpages::content()}
}
\description{
Renders a reactable table of the provided content items
}
